backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
pydantic==2.9.2
psycopg[binary]==3.2.1
python-dotenv==1.0.1
pydantic-settings==2.4.0
alembic==1.13.2
pydantic[email]>=2.7
email-validator>=2.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pytest>=8.0.0
pytest-cov>=5.0.0
httpx>=0.27.0
